03_Google_cloud_functions/requirements.txt
python-dotenv==1.0.1
google-api-core==2.19.0
google-auth==2.29.0
google-cloud-bigquery==3.23.1
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0